requirements.txt
# Flask Web Framework
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7

# Database
mysql-connector-python==9.6.0

# Blockchain
web3==7.14.1

# PDF Generation
reportlab==4.2.2

# Environment Variables
python-dotenv==1.0.0

# Image Processing (optional - for facial recognition)
opencv-python==4.8.0.76
Pillow==10.0.0
numpy==1.24.3

# Note: face-recognition requires dlib which needs CMake
# face-recognition==1.3.5